+++ ci/stage-package.js
@@ -1,5 +1,3 @@
-const fs = require('fs');
-const path = require('path');
 const util = require('./ci-util');
 
 const hotfix = process.argv[2].toLowerCase() == 'true';
@@ -13,29 +11,4 @@ if (individually) {
 } else {
     // create the tasks.zip
     util.createTasksZip();
-
-    if (hotfix) {
-        // during CI agent checks out a commit, not a branch.
-        // $(build.sourceBranch) indicates the branch name, e.g. releases/m108
-        // assumes user has checked out a branch. this is a fairly safe assumption.
-        // this code only runs during "package" and "publish" build targets, which is not typically run locally.
-        const branch = process.env.TF_BUILD ? process.env.BUILD_SOURCEBRANCH : util.run('git symbolic-ref HEAD');
-
-        const commitInfo = util.run('git log -1 --format=oneline');
-
-        // create the script
-        fs.mkdirSync(util.hotfixLayoutPath);
-        const scriptPath = path.join(util.hotfixLayoutPath, `${process.env.TASK}.ps1`);
-        const scriptContent = `
-# Hotfix created from branch: ${branch}
-# Commit: ${commitInfo}
-$ErrorActionPreference='Stop'
-Update-DistributedTaskDefinitions -TaskZip $PSScriptRoot/tasks.zip
-`;
-        fs.writeFileSync(scriptPath, scriptContent);
-
-        // link the non-aggregate tasks zip
-        const zipDestPath = path.join(util.hotfixLayoutPath, 'tasks.zip');
-        fs.linkSync(util.tasksZipPath, zipDestPath);
-    }
 }
